+++ wtr_backend_project/src/controllers/AuthController.ts
@@ -1,8 +1,10 @@
 import { NextFunction, Request, Response } from "express";
 import { authService } from "../prisma/services";
-import AuthSuccessResponse from "../types/responses/AuthSuccessResponse";
-import ResourceCreatedResponse from "../types/responses/ResourceCreatedResponse";
 import BaseController from "./abstract/BaseController";
+import {
+  AuthSuccessResponse,
+  ResourceCreatedResponse,
+} from "../types/responses";
 
 export default class AuthController extends BaseController {
   async registerUser(req: Request, res: Response, next: NextFunction) {
