+++ wtr_backend_project/src/middlewares/handleError.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import ErrorResponse from "../types/responses/ErrorResponse";
-import BaseError from "../types/errors/BaseError";
+import BaseError from "../types/errors/abstract/BaseError";
 
 export default function handleError(error:BaseError,req:Request,res:Response,next:NextFunction){
     new ErrorResponse(error.httpStatusCode,error.message).send(res)
