+++ wtr_backend_project/src/types/errors/ConflictError.ts
@@ -1,4 +1,4 @@
-import BaseError from "./BaseError";
+import BaseError from "./abstract/BaseError";
 
 export default class ConflictError extends BaseError {
   constructor(resource: string) {
