+++ wtr_backend_project/src/types/errors/abstract/BaseError.ts
