+++ wtr_backend_project/src/types/errors/index.ts
@@ -1,0 +1,4 @@
+export { default as AuthorizationError } from "./AuthorizationError";
+export { default as ConflictError } from "./ConflictError";
+export { default as NotFoundError } from "./NotFoundError";
+export { default as ValidationError } from "./ValidationError";
