+++ wtr_backend_project/src/types/responses/ErrorResponse.ts
@@ -1,4 +1,4 @@
-import BaseApiResponse from "./BaseApiResponse";
+import BaseApiResponse from "./abstract/BaseApiResponse";
 
 export default class ErrorResponse extends BaseApiResponse {
   constructor(status:number=500,message:string="Internal Server Error.") {
