+++ wtr_backend_project/src/types/responses/abstract/BaseApiResponse.ts
