+++ wtr_backend_project/src/types/responses/index.ts
@@ -1,0 +1,6 @@
+export { default as AuthSuccessResponse } from "./AuthSuccessResponse";
+export { default as DataRetrievedResponse } from "./DataRetrievedResponse";
+export { default as ErrorResponse } from "./ErrorResponse";
+export { default as ResourceCreatedResponse } from "./ResourceCreatedResponse";
+export { default as ResourceUpdatedResponse } from "./ResourceUpdatedResponse";
+export { default as ValidationErrorResponse } from "./ValidationErrorResponse";
